TankPhotoArchive_Scripts/requirements_CV.txt
opencv-python
numpy
Pillow
scikit-learn
torch
torchvision
tqdm